Fix case when reading in spot_colors, not all z-planes considered.

+++ iss/utils/morphology.py
@@ -354,6 +354,7 @@ def imfilter_coords(image: np.ndarray, kernel: np.ndarray, coords: np.ndarray, p
         else:
             image2 = np.pad(image2, pad_size, padding).astype(np.int8)
     if isinstance(padding, numbers.Number):
-        return cy_convolve(np.pad(image, pad_size, 'constant', constant_values=padding), kernel, pad_coords, image2)
+        return cy_convolve(np.pad(image, pad_size, 'constant', constant_values=padding), kernel,
+                           pad_coords.astype(np.int32), image2)
     else:
-        return cy_convolve(np.pad(image, pad_size, padding), kernel, pad_coords, image2)
+        return cy_convolve(np.pad(image, pad_size, padding), kernel, pad_coords.astype(np.int32), image2)
